+++ packages/otel/src/assemble-operation-name.ts
@@ -1,0 +1,21 @@
+import type { TelemetrySettings } from 'ai';
+
+export function assembleOperationName({
+  operationId,
+  telemetry,
+}: {
+  operationId: string;
+  telemetry?: TelemetrySettings;
+}) {
+  return {
+    // standardized operation and resource name:
+    'operation.name': `${operationId}${
+      telemetry?.functionId != null ? ` ${telemetry.functionId}` : ''
+    }`,
+    'resource.name': telemetry?.functionId,
+
+    // detailed, AI SDK specific data:
+    'ai.operationId': operationId,
+    'ai.telemetry.functionId': telemetry?.functionId,
+  };
+}