+++ packages/azure/CHANGELOG.md
@@ -1,5 +1,47 @@
 # @ai-sdk/azure
 
+## 4.0.0-beta.21
+
+### Patch Changes
+
+- c29a26f: feat(provider): add support for provider references and uploading files as supported per provider
+- Updated dependencies [c29a26f]
+  - @ai-sdk/provider-utils@5.0.0-beta.10
+  - @ai-sdk/provider@4.0.0-beta.6
+  - @ai-sdk/openai@4.0.0-beta.21
+
+## 4.0.0-beta.20
+
+### Patch Changes
+
+- 38fc777: Add AI Gateway hint to provider READMEs
+- Updated dependencies [38fc777]
+  - @ai-sdk/openai@4.0.0-beta.20
+
+## 4.0.0-beta.19
+
+### Patch Changes
+
+- Updated dependencies [2e17091]
+  - @ai-sdk/provider-utils@5.0.0-beta.9
+  - @ai-sdk/openai@4.0.0-beta.19
+
+## 4.0.0-beta.18
+
+### Patch Changes
+
+- Updated dependencies [986c6fd]
+- Updated dependencies [493295c]
+  - @ai-sdk/provider-utils@5.0.0-beta.8
+  - @ai-sdk/openai@4.0.0-beta.18
+
+## 4.0.0-beta.17
+
+### Patch Changes
+
+- Updated dependencies [817a1a6]
+  - @ai-sdk/openai@4.0.0-beta.17
+
 ## 4.0.0-beta.16
 
 ### Patch Changes